voice-receiver/requirements.txt
kafka-python=2.0.2
pyhton-dotenv=1.0.0